+++ packages/article-comments/src/utils.js
@@ -1,0 +1,15 @@
+const storeURL = {
+  gb:
+    "https://thetimes.co.uk/subscribe/digital?ILC=GB-TNL_The_Times-Conversion_Page-Homepage-2020",
+  ie:
+    "https://store.thetimes.ie/?ILC=IE-TNL_The_Times-Conversion_Page-Homepage-2020",
+  global:
+    "https://globalstore.thetimes.co.uk/?ILC=INTL-TNL_The_Times-Conversion_Page-Homepage-2020"
+};
+
+export default () => {
+  const region =
+    // eslint-disable-next-line no-undef
+    typeof window !== "undefined" && window.nuk && window.nuk.region;
+  return storeURL[region] || storeURL.gb;
+};