+++ packages/article-magazine-comment/src/article-magazine-comment.js
@@ -80,7 +80,6 @@ class ArticlePage extends Component {
       isPreview,
       swgProductId,
       additionalRelatedArticlesFlag,
-      inlineRelatedArticlesFlag,
       algoliaSearchKeys,
       latestFromSectionFlag,
       latestFromSection,
@@ -104,7 +103,6 @@ class ArticlePage extends Component {
         isPreview={isPreview}
         swgProductId={swgProductId}
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
-        inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}
         latestFromSectionFlag={latestFromSectionFlag}
         latestFromSection={latestFromSection}