+++ packages/article-magazine-standard/src/article-magazine-standard.js
@@ -70,7 +70,6 @@ class ArticlePage extends Component {
       isPreview,
       swgProductId,
       additionalRelatedArticlesFlag,
-      inlineRelatedArticlesFlag,
       algoliaSearchKeys,
       latestFromSectionFlag,
       latestFromSection,
@@ -94,7 +93,6 @@ class ArticlePage extends Component {
         isPreview={isPreview}
         swgProductId={swgProductId}
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
-        inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}
         latestFromSectionFlag={latestFromSectionFlag}
         latestFromSection={latestFromSection}