+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.64.10](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.64.9...@times-components/article-skeleton@1.64.10) (2022-01-12)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.64.9](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.64.8...@times-components/article-skeleton@1.64.9) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.64.8](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.64.7...@times-components/article-skeleton@1.64.8) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.64.7](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.64.6...@times-components/article-skeleton@1.64.7) (2022-01-07)
+
+
+### Bug Fixes
+
+* **TPD-997:** Remove inline related articles flag ([#2834](https://github.com/newsuk/times-components/issues/2834)) ([ffc332a](https://github.com/newsuk/times-components/commit/ffc332a0bf096e5436427b990a495bb3caff80ff))
+
+
+
+
+
+## [1.64.6](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.64.5...@times-components/article-skeleton@1.64.6) (2022-01-04)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 ## [1.64.5](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.64.4...@times-components/article-skeleton@1.64.5) (2022-01-04)
 
 **Note:** Version bump only for package @times-components/article-skeleton