+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.16.72](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.16.71...@times-components/article-summary@3.16.72) (2022-01-12)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
+## [3.16.71](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.16.70...@times-components/article-summary@3.16.71) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
+## [3.16.70](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.16.69...@times-components/article-summary@3.16.70) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
+## [3.16.69](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.16.68...@times-components/article-summary@3.16.69) (2022-01-07)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
+## [3.16.68](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.16.67...@times-components/article-summary@3.16.68) (2022-01-04)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
 ## [3.16.67](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.16.66...@times-components/article-summary@3.16.67) (2022-01-04)
 
 **Note:** Version bump only for package @times-components/article-summary