+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.6.26](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.6.25...@times-components/related-articles@6.6.26) (2022-01-12)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.6.25](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.6.24...@times-components/related-articles@6.6.25) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.6.24](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.6.23...@times-components/related-articles@6.6.24) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.6.23](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.6.22...@times-components/related-articles@6.6.23) (2022-01-07)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.6.22](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.6.21...@times-components/related-articles@6.6.22) (2022-01-04)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
 ## [6.6.21](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.6.20...@times-components/related-articles@6.6.21) (2022-01-04)
 
 **Note:** Version bump only for package @times-components/related-articles